Rename all.py to app.py
Alteração necessária para rodar corretamente os testes e o servidor Azure.
